Try to remove replace.
